requirements.txt
Django==5.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
psycopg2-binary==2.9.9
